long and lat

+++ src/main/java/com/shoploc/shoploc/domain/account/AccountController.java
@@ -34,8 +34,8 @@ public class AccountController {
                                                 @RequestParam ("email")String email,
                                                 @RequestParam ("roleId")Integer roleId,
                                                 @RequestParam(name="image",required=false) MultipartFile image,
-                                                @RequestParam(name="longitude",required = false) Integer longitude,
-                                                @RequestParam(name = "latitude",required = false)Integer latitude) throws InsertionFailedException, IOException {
+                                                @RequestParam(name="longitude",required = false) Double longitude,
+                                                @RequestParam(name = "latitude",required = false)Double latitude) throws InsertionFailedException, IOException {
 
         System.out.println(image+"****************************************************************");
         this.accountService.createAccount(firstname,lastname,email,roleId,image,longitude,latitude);
