+++ src/main/java/com/shoploc/shoploc/domain/store/Store.java
@@ -21,27 +21,27 @@ public class Store {
 
     private String name;
 
-    private Integer longitude;
-    private Integer latitude;
+    private Double longitude;
+    private Double latitude;
     //todo : longitude and lan
     @Lob
     private String image;
 
     private String email;
 
-    public Integer getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Integer longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 
-    public Integer getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Integer latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
